refactor: migrate province overlay demo data to seed-42 fixtures (#3847)
Replace getDebugInitGameResult in province_overlay_demo_data with lazy-loaded
committed JSON via lib/test_support seed42 loaders; test support re-exports.
